refactor: extract shared state definitions
